+++ pkg/interpreter/interpreter_integration_test.go
@@ -286,6 +286,56 @@ var integrationTests = []struct {
 		wantCategory:   "",
 		config:         interpreter.DefaultConfig(),
 	},
+	// v0.24.0 regression tests
+	{
+		name:           "slice 3index",
+		dir:            "slice_3index",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "reflect type of",
+		dir:            "reflect_type_of",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "reflect deep equal",
+		dir:            "reflect_deep_equal",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "xml marshal",
+		dir:            "xml_marshal",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "csv readall",
+		dir:            "csv_readall",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "flag parse",
+		dir:            "flag_parse",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "runtime numcpu",
+		dir:            "runtime_numcpu",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
 	// v0.23.0 regression tests
 	{
 		name:           "hex encode",